[REFACTOR] #47 : dateUtil class -> object로 변경
