+++ feature/notice/src/main/java/com/wap/wapp/feature/notice/Calendar.kt
@@ -29,8 +29,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.wap.designsystem.WappTheme
 import com.wap.designsystem.component.Loading
-import com.wap.wapp.feature.notice.DateUtil.Companion.DAYS_IN_WEEK
-import com.wap.wapp.feature.notice.DateUtil.Companion.yyyyMMddFormatter
+import com.wap.wapp.feature.notice.DateUtil.DAYS_IN_WEEK
+import com.wap.wapp.feature.notice.DateUtil.yyyyMMddFormatter
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import java.time.DayOfWeek
